Add functionality for managing editor conflicts
We now Poll GitHub for changes, and sync when there is an update available.
The entries are synced, unless a user is editing them, in which case a 'diff'
display is shown, which displays the fields that have been changed.
The GitHub repo is polled immediately before saving to see if there has been
a new commit. If there has, then the save is aborted, and the user is asked to
ensure that they have resolved any conflicts.
To facilitate these changes a new field called PS_UUID is added to all document
fragments. This field helps manage diffs and ensure that we can keep track of
fragments as they update and sync. This field is not searchable and is not
displayed in the UI.

Connects to #149

change-type: minor