Add a function to import entries from a markdown document
Fixes #201

An entry is denoted by an h2 tag (`##` in markdown), all text until the next h2
tag is considered to be part of the entry. The value of the h2 tag is used as
the entry title. The first element in the schema is treated as the entry title.
Any h3 tag (`###` in markdown) is converted to an entry field. The value for the h3
tag is used as the entry field's name and all text until the next occurring h3
tag is considered to be the value of the field.
If there is text immediately after an h2 tag, it is added to the entry using the
key Imported copy.

change-type: minor